Rename connection.js to connectionProvider.js Update client to leave the responsibility to determine connection settings type to connectionProvider Update postgres dialect aquireConnection function
