+++ lib/client.js
@@ -15,7 +15,7 @@ const TableBuilder = require('./schema/tablebuilder');
 const TableCompiler = require('./schema/tablecompiler');
 const ColumnBuilder = require('./schema/columnbuilder');
 const ColumnCompiler = require('./schema/columncompiler');
-const Connection = require('./connection');
+const ConnectionConfigProvider = require('./connectionProvider');
 
 const { Pool, TimeoutError } = require('tarn');
 const inherits = require('inherits');
@@ -55,11 +55,11 @@ function Client(config = {}) {
     this.version = config.version;
   }
   // if the type of connection settings is a function, pass it to the next layer
-  if (typeof config.connection === 'function') {
-    this.connectionSettings = config.connection;
-  } else {
-    this.connectionSettings = cloneDeep(config.connection || {});
-  }
+  // if (typeof config.connection === 'function') {
+  //   this.connectionSettings = config.connection;
+  // } else {
+  //   this.connectionSettings = cloneDeep(config.connection || {});
+  // }
   if (this.driverName && config.connection) {
     this.initializeDriver();
     if (!config.pool || (config.pool && config.pool.max !== 0)) {
@@ -127,8 +127,8 @@ Object.assign(Client.prototype, {
     return new Ref(this, ...arguments);
   },
 
-  connection() {
-    return new Connection(this);
+  connectionConfigProvider() {
+    return new ConnectionConfigProvider(this);
   },
 
   _formatQuery(sql, bindings, timeZone) {
