+++ lib/dialects/postgres/index.js
@@ -107,8 +107,10 @@ Object.assign(Client_PG.prototype, {
   acquireRawConnection() {
     const client = this;
     return new Bluebird(async function(resolver, rejecter) {
-      await client.connection().resolveConnectionConfig();
-      const connection = new client.driver.Client(client.connectionSettings);
+      const connectionSettings = await client
+        .connectionConfigProvider()
+        .resolveConnectionConfig();
+      const connection = new client.driver.Client(connectionSettings);
       connection.connect(function(err, connection) {
         if (err) {
           return rejecter(err);
